+++ lib/src/services/openpanel_http_client.dart
@@ -18,7 +18,9 @@ class OpenpanelHttpClient {
     required Logger logger,
   }) : _logger = logger;
 
-  void updateProfile({required UpdateProfilePayload payload, required Map<String, dynamic> stateProperties}) {
+  void updateProfile(
+      {required UpdateProfilePayload payload,
+      required Map<String, dynamic> stateProperties}) {
     runApiCall(() async {
       await dio.post('/profile', data: {
         ...payload.toJson(),
@@ -30,7 +32,10 @@ class OpenpanelHttpClient {
     });
   }
 
-  void increment({required String profileId, required String property, required int value}) {
+  void increment(
+      {required String profileId,
+      required String property,
+      required int value}) {
     runApiCall(() async {
       dio.post('/profile/increment', data: {
         'profileId': profileId,
@@ -40,7 +45,10 @@ class OpenpanelHttpClient {
     });
   }
 
-  void decrement({required String profileId, required String property, required int value}) {
+  void decrement(
+      {required String profileId,
+      required String property,
+      required int value}) {
     runApiCall(() async {
       dio.post('/profile/decrement', data: {
         'profileId': profileId,